pyproject.toml
[project]
name = "sql-fixer-env"
version = "0.1.0"
description = "An OpenEnv environment for fixing broken SQL queries against a SQLite database"
authors = [{name = "Akshat Jaiswal", email = "akshatjaissofficial@gmail.com"}]
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "websockets",
    "openai"
]
requires-python = ">=3.10"

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["sql_fixer_env*", "server*", "models*", "client*"]